@@ -41,43 +41,62 @@
       "@commitlint/config-conventional"
     ]
   },
-  "contributors": [
-    {
-      "name": "Robert Kieffer",
-      "email": "robert@broofa.com"
-    },
-    {
-      "name": "Christoph Tavan",
-      "email": "dev@tavan.de"
-    },
-    {
-      "name": "AJ ONeal",
-      "email": "coolaj86@gmail.com"
-    },
-    {
-      "name": "Vincent Voyer",
-      "email": "vincent@zeroload.net"
-    },
-    {
-      "name": "Roman Shtylman",
-      "email": "shtylman@gmail.com"
-    }
-  ],
   "deprecated": false,
   "description": "RFC4122 (v1, v4, and v5) UUIDs",
   "devDependencies": {
-    "@commitlint/cli": "~8.2.0",
-    "@commitlint/config-conventional": "~8.2.0",
-    "eslint": "~6.4.0",
-    "husky": "~3.0.5",
-    "mocha": "6.2.0",
-    "runmd": "1.2.1",
-    "standard-version": "7.0.0"
+    "@babel/cli": "7.8.4",
+    "@babel/core": "7.9.0",
+    "@babel/preset-env": "7.9.5",
+    "@commitlint/cli": "8.3.5",
+    "@commitlint/config-conventional": "8.3.4",
+    "@rollup/plugin-node-resolve": "7.1.3",
+    "@wdio/browserstack-service": "6.0.12",
+    "@wdio/cli": "6.0.15",
+    "@wdio/jasmine-framework": "6.0.15",
+    "@wdio/local-runner": "6.0.15",
+    "@wdio/spec-reporter": "6.0.14",
+    "@wdio/static-server-service": "6.0.13",
+    "@wdio/sync": "6.0.15",
+    "babel-eslint": "10.1.0",
+    "bundlewatch": "0.2.6",
+    "eslint": "6.8.0",
+    "eslint-config-prettier": "6.10.1",
+    "eslint-config-standard": "14.1.1",
+    "eslint-plugin-import": "2.20.2",
+    "eslint-plugin-node": "11.1.0",
+    "eslint-plugin-prettier": "3.1.3",
+    "eslint-plugin-promise": "4.2.1",
+    "eslint-plugin-standard": "4.0.1",
+    "husky": "4.2.5",
+    "jest": "25.3.0",
+    "lint-staged": "10.1.3",
+    "npm-run-all": "4.1.5",
+    "prettier": "2.0.4",
+    "rollup": "2.6.1",
+    "rollup-plugin-terser": "5.3.0",
+    "runmd": "1.3.2",
+    "standard-version": "7.1.0"
   },
+  "exports": {
+    "./package.json": "./package.json",
+    ".": {
+      "require": "./dist/index.js",
+      "import": "./wrapper.mjs"
+    }
+  },
+  "files": [
+    "CHANGELOG.md",
+    "CONTRIBUTING.md",
+    "LICENSE.md",
+    "README.md",
+    "dist",
+    "wrapper.mjs"
+  ],
   "homepage": "https://github.com/uuidjs/uuid#readme",
   "husky": {
     "hooks": {
-      "commit-msg": "commitlint -E HUSKY_GIT_PARAMS"
+      "commit-msg": "commitlint -E HUSKY_GIT_PARAMS",
+      "pre-commit": "lint-staged"
     }
   },
   "keywords": [
